Generalized the hashing function used for simhash
