+++ freediscovery/dupdet/__init__.py
@@ -5,6 +5,6 @@ from __future__ import division
 from __future__ import print_function
 
 
-from .base import SimhashDuplicates
-from .dupdet import DuplicateDetection
+from .simhash import SimhashDuplicates
+from .base import DuplicateDetection
 from .imatch import IMatchDuplicates
